ast.rs: only replace symbols in last step of reduction
Otherwise, since beta_reduce_once_recur is a recursive function,
symbols may be substituted when not needed, e.g. in

id = \x -> x
\x -> x id

Since id doesn't appear in the left side of a redex, it's not necessary
to substitute it yet.
However, if the entire expression has reduced to 'id', we do want this
symbol to be replaced; otherwise it might not be clear that there exists
a definition for this symbol.