ast.rs: Revert bug in single variable symbol lookup
Commit 4913a6d39be0dc650a92d41f80b9e6445f43aa1c introduced a bug:
if the entire expression is a single free variable it was correctly
looked up and replaced by its definition in the symbol table, if any, but
that new expression was never printed. Make last_step_beta_reduce also
return a boolean to indicate whether such substitution happened,
to decide whether to return immediately or run until the end of the
iteration in order to print the last expression.