parser.rs: fix corner case at test bound_vars_inside_multiple_lambdas
In this test, the expression used was
        \f -> (\x -> f) (\x -> f)
The last f was considered by the parser to be a free variable despite
being inside the lambda term with variable "f". The problem was caused
by this line at LineParser::parse_ast():
        self.lambda_vars.clear();
This caused all lambda variables to be removed after finalizing a
lambda term, which happened in our test after parsing the first
occurence of (\x -> f).
Fix this so only the variables inserted by the last call to
LineParser::parse_lambda() are removed. To that end, change lambda_vars
to a stack (Vec), since HashSet doesn't retain information about the
order in which elements are inserted, and make parse_lambda() also
return the number of lambda variables found.