ast.rs: adapt tests for unnecessary alpha conversion
Performing alpha conversions when not strictly necessary does not lead
to a wrong reduction, and solving it would require us to keep track of
whether the bound variable about to be replaced occurs in sub lambda
terms in get_all_lambda_vars(). This is needlessly complicated for
something that's not really a bug.